python/requirements.txt
dbr
mrz-scanner-sdk
document-scanner-sdk
opencv-python
flask